default.xml: reorder to match 'repo manifest -o'
Re-order default.xml so it matches the output from 'repo manifest -o' in
a sync'd workspace.

This allows pinned manifests to be compared more easily.

Note, 'repo manifest -o' introduces spaces on blank lines. These
are left in to stick to the exact repo output.

Based on the original commit to 96boards/oe-rpb-manifest by Ryan Harkin
<ryan.harkin@linaro.org>

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>