manifest: remove relative symlink
All along we have been using a relative symlink for setup-environment
which points outside of the project path. While we've enjoyed this
hack for several years, a recent change was in repo, and repo init now
fails with:

error.ManifestInvalidPathError: <linkfile> invalid "src": ../../.repo/manifests/setup-environment: bad component: .repo

The change in repo is this one:
https://gerrit.googlesource.com/git-repo/+/07392ed32662006c029299bc06617310cfe17957

The repo documentation states: The symlink is created at “dest”
(relative to the top of the tree) and points to the path specified by
“src” which is a path in the project.

As such, repo clearly expects that 'src' is in the project path, which
is not the cause when using ../../.repo/ path.

This patch changes our folder structure in such a way that:
* the manifest git is checked out as ./conf/ folder in the top level
directory
* setup-environment link is created by repo, as a link to
conf/setup-environment-internal

The main drawback is that it feels non standard to duplicate the whole
manifest project again (since it's already checked out in .repo)
however using .repo content also meant we were making assumptions on
implementation details of repo itself.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>

[Andrey Konovalov: adapted for lhg-oe-manifests]

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>